requirements.txt
numpy<2
matplotlib
roboticstoolbox-python
PyQt5
pandas